dodanie rejestracji oraz konwersacji w pokojach

+++ base/views.py
@@ -1,17 +1,24 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib import messages
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm, TopicForm
 from django.contrib.auth import login, logout, authenticate
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 
 def login_page(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     #standardowe sprawdzenie czy użytkownik podał dane
     if request.method == 'POST':
         #pobieramy login i hasło do zmiennych
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         #upewniamy się że użytkownik istnieje w bazie danych
@@ -30,13 +37,35 @@ def login_page(request):
         else:
             messages.error(request, 'Username OR Password does not exist!')
 
-    return render(request, 'base/login_registration.html')
+    context = {'page': page}
+    return render(request, 'base/login_registration.html', context)
 
 def logout_user(request):
 
     logout(request)
     return redirect('home')
 
+def register_user(request):
+    page = 'register'
+
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration!')
+
+
+    context = {'page': page,
+               'form': form}
+    return render(request, 'base/login_registration.html', context)
 
 def home(request):
 
@@ -62,16 +91,30 @@ def room(request, pk):
 
     room = Room.objects.get(id=pk)
 
-    context = {'room': room}
+    if request.method == 'POST':
+        body = request.POST.get("body")
+        message = Message.objects.create(body=body, room=room, user=request.user)
+        message.save()
+        return redirect('room', pk)
+
+    room_messages = room.message_set.all().order_by('created')
+
+    context = {'room': room,
+               'room_messages': room_messages}
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url="login_page")
 def create_room(request):
 
     form = RoomForm()
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
-            form.save()
+            room = form.save()
+            room.user = request.user
+            room.save()
+
             return redirect('home')
 
     context = {'form': form}
@@ -80,6 +123,9 @@ def create_room(request):
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.user:
+        return HttpResponse('You are not allowed here!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -93,10 +139,12 @@ def update_room(request, pk):
 def delete_room(request, pk):
 
     room = Room.objects.get(id=pk)
-    if request.method == 'POST':
 
-        room.delete()
-        return redirect('home')
+    if request.user == room.user:
+        if request.method == 'POST':
+
+            room.delete()
+            return redirect('home')
 
     return render(request, 'base/delete.html', {'obj': room})
 
@@ -107,6 +155,7 @@ def topic_list(request):
     content = {'topics': topics}
     return render(request, 'base/topic_list.html', content)
 
+@login_required(login_url="login_page")
 def create_topic(request):
     form = TopicForm()
